requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-editorjs==0.2.1
djangorestframework==3.12.2
gTTS==2.2.2
gunicorn==20.0.4
html2text==2020.1.16
idna==2.10
mutagen==1.45.1
oauthlib==3.1.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
pydub==0.24.1
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
